Add test for pbc
Solve issue with rotational diffusion

Hopefully resolves the issue with the Gaussian random walk on a sphere.
Just drawing the rotation angle from a normal distribution does not
take the metric of polar coordinates into account. Would need to draw
from

    `x NormalDistribution(x)`

instead. Seems not to be trivial.

So instead the orienation vector is rotated by a normal distributed
angle around a uniformly random perpendicular axis. Then another angle
is drawn, and the vector is additionally rotated around an axis
perpendicular to both the original orientation and the first axis. This
basically simulates choosing a Gaussian random vector for every
possible direction like in the flat 2D case.