Implement a solver for the magnetic dipole moment created mean magnetic field.
Part of it is a module to calculate the polarization of a distribution.

The magnetic field is calculated by convolving the magnetic dipole moment
tensor over the mean magnetic moments of the particles. To obtain the
force the vector gradient of the magnetic field is precalculated. Both
algorithm use a spectreal FFT method.